Last Commit - Finished Endpoint section. - Renamed SQL Files to make their order more logical - Fixed some things that were wrong

+++ STGeneticsTest/SQL Scripts/2. Create Table Animals.sql
@@ -8,6 +8,6 @@ CREATE TABLE Animals
 	Breed varchar(MAX) NOT NULL,
 	BirthDate date NOT NULL,
 	Sex varchar(1) NOT NULL CONSTRAINT Sex_Constraint CHECK (Sex='M' OR Sex='F'),
-	Price int NOT NULL,
+	Price float NOT NULL,
 	Status bit NOT NULL,
 	)
